Add controller y model de Candidato
Add controller y model de Candidato
